Renamed Cxbx -> gui folder
