+++ src/gui/DbgConsole.h
