+++ src/gui/DlgAbout.cpp
