+++ src/gui/DlgAbout.h
