+++ src/gui/DlgControllerConfig.cpp
