+++ src/gui/DlgEepromConfig.cpp
