+++ src/gui/DlgLoggingConfig.cpp
