+++ src/gui/DlgLoggingConfig.h
