+++ src/gui/DlgVideoConfig.cpp
