+++ src/gui/DlgVideoConfig.h
