+++ src/gui/DlgVirtualSBCFeedback.cpp
