+++ src/gui/Wnd.cpp
