+++ src/gui/Wnd.h
