+++ src/gui/WndMain.cpp
