+++ src/gui/WndMain.h
