+++ src/gui/WndSBCFeedback.cpp
